requirements.txt
django-cors-headers
django
djangorestframework
markdown
django-filter
